Retry new or modified tests based on git diff
The best time to resolve flaky test is once you wrote that test. This is
how Google looks for flakes in their monorepo. At that moment you have
the best knowledge to solve such tests. Otherwise after month of running
such test the cognitive load to fix it is very high.

This PR adds a way to discover all changed test classes with diff to
master branch and retry them until the maxRetry or to the first fail
execution.

Signed-off-by: Michał Rowicki <mirowicki@gmail.com>