Add support for Gradle 8
Gradle 8 changes the constructor for instantiating
test frameworks for JUnit 4 and JUnit Platform.

For retrying tests, we basically want to create a copy
of the original test framework, but with a new configured
test filter, which includes the tests to be retried.

This change uses the `copyWithFilters` method on the test
framework, which we hopefully introduce in Gradle 8 soon.

Signed-off-by: Daniel Clausen <dclausen@gradle.com>