Replace test class inheritence from a spec class with a common abstract spec class
With the current setup, the IDE won't be able to recognize that methods of `SpockFuncTest` can also be executed in the context of e.g. `Spock2FuncTest`. This happens because `SpockFuncTest` is not an abstract class and it can't be made one, because its own tests won't run. I introduced a common abstract base spec that is now shared by all children specs. This setup supports executing tests in the context of any child spec via IDE.

Signed-off-by: Pavlo Shevchenko <pshevchenko@gradle.com>