Add special handling for Gradle 5.0
When using Gradle 5.0 in combination with JUnit's suite runner
or when using @Nested test classes, we lack test events for the
intermediate classes.

This leads to a problem when an intermediate class is configured
for class-level retry.

In order to fix this, we manually remove specific entries from
the failed tests at the end of a round. But we only do this
for Gradle 5.0, as test reporting for newer version does not
exhibit this problem.

Signed-off-by: Daniel Clausen <dclausen@gradle.com>